Refactor `@netlify/config` debugging
