+++ packages/config/src/main.js
@@ -9,7 +9,7 @@ const { mergeDefaultConfig } = require('./default_config')
 const { throwError } = require('./error')
 const { handleFiles } = require('./files')
 const { cleanupConfig } = require('./log/cleanup')
-const { logDefaultConfig, logResult } = require('./log/options')
+const { logDefaultConfig, logResult } = require('./log/main')
 const { normalizeConfig } = require('./normalize')
 const { addDefaultOpts, normalizeOpts } = require('./options/main')
 const { parseConfig } = require('./parse')
